// src/app/Modules/rrhh/model/yfaseplazv.dto.ts
// Generated by https://quicktype.io

import { Ydotaplazv } from "./ydotaplazv.dto";
import { Yfaseplaz } from "./yfaseplaz.dto";
import { YplazavDTO } from "./yplazav.dto";

export class YFasePlazVDTO {
    codifoad:   string;
    codisiad:   number;
    codisper:   string;
    expesitu:   string;
    fexpsitu:   Date | null;
    ffinvers:   Date | null;
    finivers:   Date | null;
    id:         YFasePlazVDTOID;
    jornredu:   string;
    motisitu:   string;
    obsefase:   string;
    porcdedi:   number;
    ydotaplazv: Ydotaplazv;
    yfaseplaz:  Yfaseplaz;
    yplazav:    YplazavDTO;

    constructor() {
        this.codifoad=   '';
        this.codisiad=   -1;
        this.codisper=   '';
        this.expesitu=   '';
        this.fexpsitu=   null;
        this.ffinvers=   null;
        this.finivers=   null;
        this.id=         new YFasePlazVDTOID();
        this.jornredu=   '';
        this.motisitu=   '';
        this.obsefase=   '';
        this.porcdedi=   100;
        this.ydotaplazv= new Ydotaplazv();
        this.yfaseplaz=  new Yfaseplaz();
        this.yplazav=    new YplazavDTO();
    
    }
}

export class YFasePlazVDTOID {
    codienti:  number;
    codiplaz:  number;
    coditpla:  number;
    natuplaz:  string;
    numefase:  number;
    versfase: number;

    constructor() {
        this.codienti=  0;
        this.codiplaz=  0;
        this.coditpla=  0;
        this.natuplaz=  '';
        this.numefase=  0;
        this.versfase= 0;
    }
}