// src/app/Modules/rrhh/model/yplaza.dto.ts
// Generated by https://quicktype.io

export class YPlazaDTO {
    desctpla: string;
    fechacre: Date | null;
    id:       YPlazaDTOID;
    constructor() {
        this.desctpla='';
        this.fechacre=null;
        this.id=new YPlazaDTOID();
    }
}

export class YPlazaDTOID {
    codienti: number;
    coditpla: number;
    natuplaz: string;
    constructor() {
        this.codienti=0;
        this.coditpla=0;
        this.natuplaz='';
    }
}