// src/app/Modules/rrhh/model/entitat.dto.ts
// Generated by https://quicktype.io

export interface EntitatDTO {
    ape1r190:     string;
    ape2r190:     string;
    bic:          string;
    cargor190:    string;
    clavdomi:     string;
    clave:        string;
    codienti:     number;
    codifirma:    number;
    codimage:     number;
    cposdent:     number;
    descdent:     string;
    digifent:     string;
    digssent:     number;
    direnti:      string;
    emailr190:    string;
    iban:         string;
    letifent:     string;
    nifr190:      string;
    nombenti:     string;
    nombr190:     string;
    nresenti:     string;
    numedent:     number;
    numetent:     number;
    numifent:     number;
    numssent:     number;
    planigualdad: string;
    preftent:     number;
    prossent:     number;
    pswdenti:     string;
    repcargo:     string;
    repdni:       string;
    repfirmaCtr:  number;
    repnombre:    string;
    stActualiza:  string;
    teler190:     number;
    zonageo1:     number;
    zonageo2:     number;
}